Add Watson directory and files
This directory pertains to the IBM Watson Chatbot as well as the available Google Cloud APIs for text-to-speech and speech-to-text. The request file contains the functionality necessary to communicate with our web server and make a POST request using a provided text string as an intent for IBM Watson. Watson then returns a text string response. The text_to_speech file contains a function for Google Cloud TTS. The function receives the returned IBM Watson text response, converts it into an audio file, and plays it back under the file name: output.mp3. The text_to_speech function also allows for custom Google Cloud voice configuration.